Add LogProblem Solution
Problem statement: A logarithm is kind of like reverse exponents. There is a base and a number in a logarithm. The point of a logarithm is to find out what power you have to raise the base to get the number next to the base.
Problem Link: https://edabit.com/challenge/o7LPd9dAE5x9k7zFj